pyproject.toml
[project]
name = "showyourwork2"
description = "A workflow for open-source scientific articles"
authors = [
    { name = "Rodrigo Luger", email = "rodluger@gmail.com" },
    { name = "Dan Foreman-Mackey", email = "foreman.mackey@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT License" }
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
]
dynamic = ["version"]
dependencies = [
    "requests",
    "pyyaml",
    "jsonschema",
    "snakemake>=7.15.2", # TODO(dfm): figure out what minimum version we _actually_ need
    "click",
    "ensureconda",
]

[project.scripts]
showyourwork2 = "showyourwork2.cli:main"

[project.urls]
"Homepage" = "https://show-your.work"
"Source" = "https://github.com/showyourwork/showyourwork"
"Bug Tracker" = "https://github.com/showyourwork/showyourwork/issues"

[project.optional-dependencies]
test = ["pytest", "flask"]
examples = []
docs = [
    "requests",
    "sphinx-book-theme>=0.1.3",
    "jinja2>=2.11.3",
    "sphinxcontrib-programoutput>=0.17",
]
dev = [
    "pre-commit",
    "black",
    "ruff",
    "mypy",
    "nox",
    "types-PyYAML",
    "types-requests",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/showyourwork2/version.py"

[tool.black]
target-version = ["py38", "py39"]
line-length = 88

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["F", "I", "E", "W", "YTT", "B", "Q", "PLE", "PLR", "PLW"]
ignore = ["PLR0913"]
exclude = []

[tool.ruff.isort]
known-first-party = ["showyourwork2"]
